+++ models/test.rb
@@ -7,19 +7,27 @@ class Test < ActiveRecord::Base
 
   attr_accessible :input, :output
 
-  def verify(solution, file)
-    input_filename = "temp/test_#{id}.txt"
-    response_filename = "temp/response_#{id}.txt"
-    File.open(input_filename, 'w+') { |f| f.write(self.input) }
+  def verify(solution, file, compilation_status)
+    if compilation_status
+      input_filename = "temp/test_#{id}.txt"
+      response_filename = "temp/response_#{id}.txt"
+      File.open(input_filename, 'w+') { |f| f.write(self.input) }
 
-    system("#{file} < #{input_filename} > #{response_filename}")
+      if system("#{file} < #{input_filename} > #{response_filename}")
+        file_output = File.read(response_filename)
+        FileUtils.rm(response_filename)
+        FileUtils.rm(input_filename)
 
-    file_output = File.read(response_filename)
-    FileUtils.rm(response_filename)
-    FileUtils.rm(input_filename)
-
-    create_solution_test(solution, file_output)
-    file_output == self.output
+        create_solution_test(solution, file_output)
+        file_output == self.output
+      else
+        create_solution_test(solution, "Runtime error")
+        return false
+      end
+    else
+      create_solution_test(solution, "Compilation Error")
+      return false
+    end
   end
 
   def create_solution_test(solution, file_output)
